feat: Update book using API
Use of the update method from the BooksAPI to update the bookshelf of the
book.

ToDo: Update state or rerender when the book is updUpdate book using API

Use of the update method from the BooksAPI to update the bookshelf of the
book.

ToDo: Update state or rerender when the book is updated